+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.63] - 2024-12-19
+
+### Added
+- Person detail: For deceased people, shows † next to their name
+- Person detail: Displays death date and age at death instead of current age for deceased people
+- Person detail: "Died" date type now displays † as its icon
+
 ## [1.0.62] - 2024-12-19
 
 ### Changed